+++ src/common/src/cast/mod.rs
@@ -12,369 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use chrono::{NaiveDate, NaiveDateTime, NaiveTime};
-use itertools::Itertools;
-use snafu::{OptionExt, ResultExt, Snafu};
-use speedate::{Date as SpeedDate, DateTime as SpeedDateTime, Time as SpeedTime};
-use strum::EnumMessage;
+mod bytea;
+mod date_time;
 
-use crate::types::{Date, Time, Timestamp, Timestamptz};
-
-type Result<T> = std::result::Result<T, CastError>;
-
-#[derive(Snafu, Debug)]
-#[snafu(display("{}", inner.get_documentation().unwrap()))]
-pub struct SpeedateParseError {
-    inner: speedate::ParseError,
-}
-
-impl From<speedate::ParseError> for SpeedateParseError {
-    fn from(value: speedate::ParseError) -> Self {
-        Self { inner: value }
-    }
-}
-
-// impl<E> std::error::Error for BaseError<E> where E: std::fmt {
-// }
-
-#[derive(Snafu, Debug)]
-pub enum CastError {
-    #[snafu(display("can't cast `{from}` to date (expected format is YYYY-MM-DD)"))]
-    DateParse {
-        from: Box<str>,
-        #[snafu(source(from(speedate::ParseError, Into::into)), provide(false))]
-        source: SpeedateParseError,
-    },
-
-    #[snafu(display(
-        "can't cast `{from}` to time (expected format is HH:MM:SS[.D+{{up to 6 digits}}][Z] or HH:MM)",
-    ))]
-    TimeParse {
-        from: Box<str>,
-        #[snafu(source(from(speedate::ParseError, Into::into)), provide(false))]
-        source: SpeedateParseError,
-    },
-
-    #[snafu(display(
-        "can't cast `{from}` to timestamp (expected format is YYYY-MM-DD HH:MM:SS[.D+{{up to 6 digits}}] or YYYY-MM-DD HH:MM or YYYY-MM-DD or ISO 8601 format)",
-    ))]
-    TimestampParse {
-        from: Box<str>,
-        #[snafu(source(from(speedate::ParseError, Into::into)), provide(false))]
-        source: SpeedateParseError,
-    },
-
-    #[snafu(display("can't cast integer `{from}` to timestamp"))]
-    TimestampFromInt { from: i64 },
-
-    #[snafu(display("invalid hex digit `{from}`"))]
-    InvalidHexDigit { from: char },
-
-    #[snafu(display("invalid format (`{from}`) for bytea: {message}"))]
-    InvalidBytea { from: Box<str>, message: Box<str> },
-}
-
-pub const PARSE_ERROR_STR_TO_TIMESTAMP: &str = "";
-pub const PARSE_ERROR_STR_TO_TIME: &str = "Can't cast string to time ()";
-pub const PARSE_ERROR_STR_TO_DATE: &str =
-    "Can't cast string to date (expected format is YYYY-MM-DD)";
-pub const PARSE_ERROR_STR_TO_BYTEA: &str = "Invalid Bytea syntax";
-
-const ERROR_INT_TO_TIMESTAMP: &str = "Can't cast negative integer to timestamp";
-
-pub fn str_to_date(elem: &str) -> Result<Date> {
-    Ok(Date::new(parse_naive_date(elem)?))
-}
-
-pub fn str_to_time(elem: &str) -> Result<Time> {
-    Ok(Time::new(parse_naive_time(elem)?))
-}
-
-pub fn str_to_timestamp(elem: &str) -> Result<Timestamp> {
-    Ok(Timestamp::new(parse_naive_datetime(elem)?))
-}
-
-#[inline]
-pub fn parse_naive_date(s: &str) -> Result<NaiveDate> {
-    let res = SpeedDate::parse_str(s).context(DateParseSnafu { from: s })?;
-    Ok(Date::from_ymd_uncheck(res.year as i32, res.month as u32, res.day as u32).0)
-}
-
-#[inline]
-pub fn parse_naive_time(s: &str) -> Result<NaiveTime> {
-    let s_without_zone = s.trim_end_matches('Z');
-    let res = SpeedTime::parse_str(s_without_zone).context(TimeParseSnafu { from: s })?;
-    Ok(Time::from_hms_micro_uncheck(
-        res.hour as u32,
-        res.minute as u32,
-        res.second as u32,
-        res.microsecond,
-    )
-    .0)
-}
-
-#[inline]
-pub fn parse_naive_datetime(s: &str) -> Result<NaiveDateTime> {
-    if let Ok(res) = SpeedDateTime::parse_str(s) {
-        Ok(Date::from_ymd_uncheck(
-            res.date.year as i32,
-            res.date.month as u32,
-            res.date.day as u32,
-        )
-        .and_hms_micro_uncheck(
-            res.time.hour as u32,
-            res.time.minute as u32,
-            res.time.second as u32,
-            res.time.microsecond,
-        )
-        .0)
-    } else {
-        let res = SpeedDate::parse_str(s).context(TimestampParseSnafu { from: s })?;
-        Ok(
-            Date::from_ymd_uncheck(res.year as i32, res.month as u32, res.day as u32)
-                .and_hms_micro_uncheck(0, 0, 0, 0)
-                .0,
-        )
-    }
-}
-
-/// Converts UNIX epoch time to timestamptz.
-///
-/// The input UNIX epoch time is interpreted as follows:
-///
-/// - [0, 1e11) are assumed to be in seconds.
-/// - [1e11, 1e14) are assumed to be in milliseconds.
-/// - [1e14, 1e17) are assumed to be in microseconds.
-/// - [1e17, upper) are assumed to be in nanoseconds.
-///
-/// This would cause no problem for timestamp in [1973-03-03 09:46:40, 5138-11-16 09:46:40).
-#[inline]
-pub fn i64_to_timestamptz(t: i64) -> Result<Timestamptz> {
-    const E11: i64 = 100_000_000_000;
-    const E14: i64 = 100_000_000_000_000;
-    const E17: i64 = 100_000_000_000_000_000;
-    match t {
-        0..E11 => Ok(Timestamptz::from_secs(t).unwrap()), // s
-        E11..E14 => Ok(Timestamptz::from_millis(t).unwrap()), // ms
-        E14..E17 => Ok(Timestamptz::from_micros(t)),      // us
-        E17.. => Ok(Timestamptz::from_micros(t / 1000)),  // ns
-        _ => TimestampFromIntSnafu { from: t }.fail(),
-    }
-}
-
-/// Converts UNIX epoch time to timestamp.
-///
-/// The input UNIX epoch time is interpreted as follows:
-///
-/// - [0, 1e11) are assumed to be in seconds.
-/// - [1e11, 1e14) are assumed to be in milliseconds.
-/// - [1e14, 1e17) are assumed to be in microseconds.
-/// - [1e17, upper) are assumed to be in nanoseconds.
-///
-/// This would cause no problem for timestamp in [1973-03-03 09:46:40, 5138-11-16 09:46:40).
-///
-/// # Example
-/// ```
-/// # use risingwave_common::cast::i64_to_timestamp;
-/// assert_eq!(
-///     i64_to_timestamp(1_666_666_666).unwrap().to_string(),
-///     "2022-10-25 02:57:46"
-/// );
-/// assert_eq!(
-///     i64_to_timestamp(1_666_666_666_666).unwrap().to_string(),
-///     "2022-10-25 02:57:46.666"
-/// );
-/// assert_eq!(
-///     i64_to_timestamp(1_666_666_666_666_666).unwrap().to_string(),
-///     "2022-10-25 02:57:46.666666"
-/// );
-/// assert_eq!(
-///     i64_to_timestamp(1_666_666_666_666_666_666)
-///         .unwrap()
-///         .to_string(),
-///     // note that we only support microseconds precision
-///     "2022-10-25 02:57:46.666666"
-/// );
-/// ```
-#[inline]
-pub fn i64_to_timestamp(t: i64) -> Result<Timestamp> {
-    let tz = i64_to_timestamptz(t)?;
-    Ok(Timestamp::from_timestamp_uncheck(
-        tz.timestamp(),
-        tz.timestamp_subsec_nanos(),
-    ))
-}
-
-/// Refer to PostgreSQL's implementation <https://github.com/postgres/postgres/blob/5cb54fc310fb84287cbdc74533f3420490a2f63a/src/backend/utils/adt/varlena.c#L276-L288>
-pub fn str_to_bytea(elem: &str) -> Result<Box<[u8]>> {
-    if let Some(remainder) = elem.strip_prefix(r"\x") {
-        Ok(parse_bytes_hex(remainder)?.into())
-    } else {
-        Ok(parse_bytes_traditional(elem)?.into())
-    }
-}
-
-/// Ref: <https://docs.rs/hex/0.4.3/src/hex/lib.rs.html#175-185>
-fn get_hex(c: u8) -> Result<u8> {
-    match c {
-        b'A'..=b'F' => Ok(c - b'A' + 10),
-        b'a'..=b'f' => Ok(c - b'a' + 10),
-        b'0'..=b'9' => Ok(c - b'0'),
-        _ => InvalidHexDigitSnafu { from: c as char }.fail(),
-    }
-}
-
-/// Refer to <https://www.postgresql.org/docs/current/datatype-binary.html#id-1.5.7.12.10> for specification.
-pub fn parse_bytes_hex(s: &str) -> Result<Vec<u8>> {
-    let mut res = Vec::with_capacity(s.len() / 2);
-
-    let mut bytes = s.bytes();
-    while let Some(c) = bytes.next() {
-        // white spaces are tolerated
-        if c == b' ' || c == b'\n' || c == b'\t' || c == b'\r' {
-            continue;
-        }
-        let v1 = get_hex(c)?;
-
-        let c = bytes.next().context(InvalidByteaSnafu {
-            from: s,
-            message: "odd number of digits in hex format",
-        })?;
-
-        let v2 = get_hex(c)?;
-        res.push((v1 << 4) | v2);
-    }
-
-    Ok(res)
-}
-
-/// Refer to <https://www.postgresql.org/docs/current/datatype-binary.html#id-1.5.7.12.10> for specification.
-pub fn parse_bytes_traditional(s: &str) -> Result<Vec<u8>> {
-    let context = InvalidByteaSnafu {
-        from: s,
-        message: "invalid input syntax for escape format",
-    };
-
-    let mut bytes = s.bytes();
-
-    let mut res = Vec::new();
-    while let Some(b) = bytes.next() {
-        if b != b'\\' {
-            res.push(b);
-        } else {
-            match bytes.next() {
-                Some(b'\\') => {
-                    res.push(b'\\');
-                }
-                Some(b1 @ b'0'..=b'3') => match bytes.next_tuple() {
-                    Some((b2 @ b'0'..=b'7', b3 @ b'0'..=b'7')) => {
-                        res.push(((b1 - b'0') << 6) + ((b2 - b'0') << 3) + (b3 - b'0'));
-                    }
-                    _ => {
-                        // one backslash, not followed by another or ### valid octal
-                        return context.fail();
-                    }
-                },
-                _ => {
-                    // one backslash, not followed by another or ### valid octal
-                    return context.fail();
-                }
-            }
-        }
-    }
-
-    Ok(res)
-}
-
-#[cfg(test)]
-mod tests {
-    use std::assert_matches::assert_matches;
-
-    use super::*;
-
-    #[test]
-    fn parse_str() {
-        str_to_timestamp("1999-01-08 04:02").unwrap();
-        str_to_timestamp("1999-01-08 04:05:06").unwrap();
-        assert_eq!(
-            str_to_timestamp("2022-08-03T10:34:02Z").unwrap(),
-            str_to_timestamp("2022-08-03 10:34:02").unwrap()
-        );
-        str_to_date("1999-01-08").unwrap();
-        str_to_time("04:05").unwrap();
-        str_to_time("04:05:06").unwrap();
-
-        assert_matches!(
-            str_to_timestamp("1999-01-08 04:05:06AA").unwrap_err(),
-            CastError::TimestampParse { .. }
-        );
-        assert_matches!(
-            str_to_date("1999-01-08AA").unwrap_err(),
-            CastError::DateParse { .. }
-        );
-        assert_matches!(
-            str_to_time("AA04:05:06").unwrap_err(),
-            CastError::TimeParse { .. }
-        );
-    }
-
-    #[test]
-    fn test_bytea() {
-        use crate::types::ToText;
-        assert_eq!(str_to_bytea("fgo").unwrap().as_ref().to_text(), r"\x66676f");
-        assert_eq!(
-            str_to_bytea(r"\xDeadBeef").unwrap().as_ref().to_text(),
-            r"\xdeadbeef"
-        );
-        assert_eq!(
-            str_to_bytea("12CD").unwrap().as_ref().to_text(),
-            r"\x31324344"
-        );
-        assert_eq!(
-            str_to_bytea("1234").unwrap().as_ref().to_text(),
-            r"\x31323334"
-        );
-        assert_eq!(
-            str_to_bytea(r"\x12CD").unwrap().as_ref().to_text(),
-            r"\x12cd"
-        );
-        assert_eq!(
-            str_to_bytea(r"\x De Ad Be Ef ").unwrap().as_ref().to_text(),
-            r"\xdeadbeef"
-        );
-        assert_eq!(
-            str_to_bytea("x De Ad Be Ef ").unwrap().as_ref().to_text(),
-            r"\x7820446520416420426520456620"
-        );
-        assert_eq!(
-            str_to_bytea(r"De\\123dBeEf").unwrap().as_ref().to_text(),
-            r"\x44655c3132336442654566"
-        );
-        assert_eq!(
-            str_to_bytea(r"De\123dBeEf").unwrap().as_ref().to_text(),
-            r"\x4465536442654566"
-        );
-        assert_eq!(
-            str_to_bytea(r"De\\000dBeEf").unwrap().as_ref().to_text(),
-            r"\x44655c3030306442654566"
-        );
-
-        assert_eq!(str_to_bytea(r"\123").unwrap().as_ref().to_text(), r"\x53");
-        assert_eq!(str_to_bytea(r"\\").unwrap().as_ref().to_text(), r"\x5c");
-        assert_eq!(
-            str_to_bytea(r"123").unwrap().as_ref().to_text(),
-            r"\x313233"
-        );
-        assert_eq!(
-            str_to_bytea(r"\\123").unwrap().as_ref().to_text(),
-            r"\x5c313233"
-        );
-
-        assert!(str_to_bytea(r"\1").is_err());
-        assert!(str_to_bytea(r"\12").is_err());
-        assert!(str_to_bytea(r"\400").is_err());
-        assert!(str_to_bytea(r"\378").is_err());
-        assert!(str_to_bytea(r"\387").is_err());
-        assert!(str_to_bytea(r"\377").is_ok());
-    }
-}
+pub use bytea::*;
+pub use date_time::*;
